08.12 - final refactorings (step 8 - final cleanup) - create an admin and common package - move the AdminController to the admin package. move the common exeptions and dtos and the LoggingFilter to the common package
