+++ src/main/java/com/jparkkennaby/store/common/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.controllers;
+package com.jparkkennaby.store.common;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -10,7 +10,6 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-import com.jparkkennaby.store.dtos.ErrorDto;
 import com.jparkkennaby.store.exceptions.MaxTableRecordLimitReached;
 
 @ControllerAdvice
