+++ src/main/java/com/jparkkennaby/store/common/LoggingFilter.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.filters;
+package com.jparkkennaby.store.common;
 
 import java.io.IOException;
